extract components from TruncatedText
