+++ src/TruncatedText/TruncatedText.tsx
@@ -1,8 +1,9 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
-import { Tooltip } from "../Tooltip";
 import { Text } from "../Type";
 import { TooltipProps } from "../Tooltip/Tooltip";
+import TruncatedTextFillWidth from "./TruncatedTextFillWidth";
+import TruncatedTextMaxCharacters from "./TruncatedTextMaxCharacters";
 
 type TruncatedTextProps = {
   children?: string;
@@ -15,111 +16,6 @@ type TruncatedTextProps = {
   tooltipProps?: TooltipProps;
 };
 
-type MaybeTooltipProps = TooltipProps & {
-  showTooltip?: boolean;
-};
-
-const MaybeTooltip = ({
-  children,
-  showTooltip,
-  ...props
-}: MaybeTooltipProps) => {
-  return showTooltip ? (
-    <Tooltip {...props}>{children}</Tooltip>
-  ) : (
-    <>{children}</>
-  );
-};
-
-MaybeTooltip.propTypes = {
-  children: PropTypes.node,
-  showTooltip: PropTypes.bool,
-};
-
-MaybeTooltip.defaultProps = {
-  children: "",
-  showTooltip: true,
-};
-
-const TruncatedTextFillWidth = ({
-  element,
-  showTooltip,
-  tooltipProps,
-  children,
-  "data-testid": testId,
-  ...props
-}: TruncatedTextProps) => {
-  const [hasOverflowText, setHasOverflowText] = useState(false);
-  const hasTooltip = showTooltip && hasOverflowText;
-  const updateOverflow = (e) => {
-    const { scrollWidth, clientWidth } = e.target;
-    if (!hasOverflowText && scrollWidth > clientWidth) {
-      setHasOverflowText(true);
-    }
-  };
-  return (
-    <MaybeTooltip
-      showTooltip={hasTooltip}
-      tooltip={children}
-      defaultOpen
-      {...tooltipProps}
-    >
-      <Text
-        as={element.type}
-        cursor={hasTooltip ? "pointer" : "initial"}
-        onMouseEnter={updateOverflow}
-        width="100%"
-        whiteSpace="nowrap"
-        overflow="hidden"
-        textOverflow="ellipsis"
-        data-testid={testId}
-        {...element.props}
-        {...props}
-      >
-        {children}
-      </Text>
-    </MaybeTooltip>
-  );
-};
-
-const TruncatedTextMaxCharacters = ({
-  children,
-  element,
-  indicator,
-  maxCharacters,
-  showTooltip,
-  tooltipProps,
-  "data-testid": testId,
-  ...props
-}: TruncatedTextProps) => {
-  const innerText = children ?? "";
-  const requiresTruncation = innerText.length > maxCharacters;
-
-  const truncatedText = requiresTruncation
-    ? innerText.slice(0, maxCharacters) + indicator
-    : innerText;
-  const hasTooltip = showTooltip && requiresTruncation;
-
-  return (
-    <MaybeTooltip
-      showTooltip={hasTooltip}
-      tooltip={innerText}
-      {...tooltipProps}
-    >
-      <Text
-        as={element.type}
-        cursor={hasTooltip ? "pointer" : "initial"}
-        width="fit-content"
-        data-testid={testId}
-        {...element.props}
-        {...props}
-      >
-        {truncatedText}
-      </Text>
-    </MaybeTooltip>
-  );
-};
-
 const TruncatedText = ({ fullWidth, children, ...props }: TruncatedTextProps) =>
   fullWidth ? (
     <TruncatedTextFillWidth {...props}>{children}</TruncatedTextFillWidth>
@@ -127,7 +23,7 @@ const TruncatedText = ({ fullWidth, children, ...props }: TruncatedTextProps) =>
     <TruncatedTextMaxCharacters {...props}>
       {children}
     </TruncatedTextMaxCharacters>
-  );
+);
 
 TruncatedText.propTypes = {
   children: PropTypes.node,
